pyproject.toml
[tool.poetry]
name = "drf-exam"
version = "0.1.0"
description = ""
authors = ["Ben <ben@ahnlab.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
